refactor: extract stream demo from AgentWorkspace
