+++ src/components/agent/AgentWorkspace.tsx
@@ -306,6 +306,7 @@ import {
 } from "../../lib/agent-chat-gallery";
 import { attachScrollThumbFade } from "../../lib/scroll-thumb-fade";
 import type { AgentWorkspaceProps } from "./agent-workspace-types";
+import { useAgentStreamDemo } from "./hooks/use-agent-stream-demo";
 import { useAgentSteerDemo } from "./hooks/use-agent-steer-demo";
 import { createCapabilityActions } from "./capability-actions";
 import { createSessionTitleActions } from "./session-title-actions";
@@ -4786,73 +4787,9 @@ export function AgentWorkspace({
   // scope): pick up the desired state on mount and follow live toggles via the
   // window event. Feeds the gallery timeline an append-only running text part
   // in irregular chunks, like a real provider stream.
-  useEffect(() => {
-    if (!import.meta.env.DEV) return;
-    let timer: number | null = null;
-    const stop = () => {
-      if (timer !== null) {
-        window.clearInterval(timer);
-        timer = null;
-      }
-    };
-    const TICK_MS = 90;
-    // Irregular-but-deterministic chunk scaling so the replay has the uneven
-    // cadence of a real provider without churn between runs.
-    const JITTER = [0.4, 1.7, 0.2, 2.3, 1, 0.6, 1.8];
-    const apply = ({ show, charsPerSecond }: AgentStreamDemoDetail) => {
-      stop();
-      if (!show) {
-        setGallerySections((prev) =>
-          prev?.[0]?.label === STREAM_DEMO_SECTION_LABEL ? null : prev,
-        );
-        return;
-      }
-      const text = SAMPLE_MARKDOWN;
-      let at = 0;
-      let tick = 0;
-      let restTicks = 0;
-      const seed = (end: number, status: "running" | "complete") =>
-        setGallerySections([
-          {
-            label: STREAM_DEMO_SECTION_LABEL,
-            description: `Replaying ~${charsPerSecond} chars/s on loop. __streamDemo(false) stops.`,
-            turns: [
-              {
-                id: "gallery:stream-demo",
-                role: "assistant",
-                createdAt: "2026-06-09T12:00:00.000Z",
-                status,
-                parts: [{ type: "text", text: text.slice(0, end), status }],
-              },
-            ],
-          },
-        ]);
-      seed(0, "running");
-      timer = window.setInterval(() => {
-        if (restTicks > 0) {
-          restTicks -= 1;
-          if (restTicks === 0) {
-            at = 0;
-            seed(0, "running");
-          }
-          return;
-        }
-        tick += 1;
-        const step = Math.round(((charsPerSecond * TICK_MS) / 1000) * JITTER[tick % JITTER.length]);
-        at = Math.min(text.length, at + Math.max(1, step));
-        const done = at >= text.length;
-        seed(at, done ? "complete" : "running");
-        if (done) restTicks = Math.round(2000 / TICK_MS);
-      }, TICK_MS);
-    };
-    apply(streamDemoDesired);
-    const onDemo = (event: Event) => apply((event as CustomEvent<AgentStreamDemoDetail>).detail);
-    window.addEventListener(AGENT_STREAM_DEMO_EVENT, onDemo);
-    return () => {
-      window.removeEventListener(AGENT_STREAM_DEMO_EVENT, onDemo);
-      stop();
-    };
-  }, []);
+  useAgentStreamDemo({
+    setGallerySections,
+  });
 
   // Reopen the steer queue whenever the open session changes — collapsing it
   // is a per-session, per-glance affordance, not a sticky mode.
