+++ src/components/agent/hooks/use-agent-stream-demo-types.ts
@@ -1,0 +1,6 @@
+import { type AgentChatGallerySection } from "../../../lib/agent-chat-gallery";
+import type * as React from "react";
+
+export type useAgentStreamDemoDependencies = {
+  setGallerySections: React.Dispatch<React.SetStateAction<AgentChatGallerySection[] | null>>;
+};
